+++ src/span.h
@@ -60,7 +60,7 @@ struct span_runtime {
 typedef uint64_t span_bytes;
 
 /* Convert offset to pointer to span. offset must be 8-bytes aligned. */
-span_bytes *span_offset_to_span_ptr(struct pmemstream *stream, size_t offset);
+span_bytes *span_offset_to_span_ptr(struct pmemstream *stream, uint64_t offset);
 
 /* Those functions create appropriate span at specified offset. offset must be 8-bytes aligned. */
 void span_create_empty(struct pmemstream *stream, uint64_t offset, size_t data_size);
